+++ gamevault/Helper/Web/WebExceptionHelper.cs
@@ -17,7 +17,7 @@ namespace gamevault.Helper
             string errMessage = string.Empty;
             try
             {
-                var resp = new StreamReader(ex.Response.GetResponseStream()).ReadToEnd();
+                var resp = new StreamReader(ex.Response?.GetResponseStream()).ReadToEnd();
                 JsonObject obj = JsonNode.Parse(resp).AsObject();
                 errMessage = obj["message"].ToString().Replace("\n", " ").Replace("\r", "");
             }