+++ gamevault/UserControls/SettingsUserControl.xaml.cs
@@ -17,6 +17,8 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Collections.Generic;
 using System.Windows.Markup;
+using gamevault.Helper.Integrations;
+using AngleSharp.Dom;
 
 namespace gamevault.UserControls
 {
@@ -248,6 +250,8 @@ namespace gamevault.UserControls
 
             if (((ThemeItem)((ComboBox)sender).SelectionBoxItem).Path == selectedTheme.Path)
                 return;
+
+            ViewModel.IsCommunityThemeSelected = File.Exists(selectedTheme?.Path);
             if (selectedTheme.IsPlus == true && ViewModel.License.IsActive() == false)
             {
                 ((ComboBox)sender).SelectedItem = (ThemeItem)((ComboBox)sender).SelectionBoxItem;
@@ -327,6 +331,7 @@ namespace gamevault.UserControls
                 {
                     uiCbTheme.SelectedIndex = 0;
                 }
+                ViewModel.IsCommunityThemeSelected = File.Exists(((ThemeItem)uiCbTheme.SelectedItem)?.Path);
             }
             catch { uiCbTheme.SelectedIndex = 0; }
         }
@@ -449,6 +454,19 @@ namespace gamevault.UserControls
             }
             ((FrameworkElement)sender).IsEnabled = true;
         }
+        private void UninstallTheme_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                if (File.Exists(((ThemeItem)uiCbTheme.SelectedItem)?.Path))
+                {
+                    File.Delete(((ThemeItem)uiCbTheme.SelectedItem).Path);
+                    uiCbTheme.SelectedIndex = 0;
+                    LoadThemes();
+                }
+            }
+            catch { }
+        }
         #endregion
         private void Hyperlink_RequestNavigate(object sender, System.Windows.Navigation.RequestNavigateEventArgs e)
         {
@@ -523,6 +541,28 @@ namespace gamevault.UserControls
                 ViewModel.DevModeEnabled = true;
             }
         }
+        private void RemoveCustomCloudSaveManifest_Click(object sender, RoutedEventArgs e)
+        {
+            int index = ViewModel.CustomCloudSaveManifests.IndexOf(((LudusaviManifestEntry)((FrameworkElement)sender).DataContext));
+            if (index >= 0)
+            {
+                ViewModel.CustomCloudSaveManifests.RemoveAt(index);
+            }
+        }
+        private void AddCustomCloudSaveManifest_Click(object sender, RoutedEventArgs e)
+        {
+            ViewModel.CustomCloudSaveManifests.Add(new LudusaviManifestEntry());
+        }
 
+        private void SaveCustomCloudSaveManifests_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                string result = string.Join(";", ViewModel.CustomCloudSaveManifests.Where(entry => !string.IsNullOrWhiteSpace(entry.Uri)).Select(entry => entry.Uri));
+                Preferences.Set(AppConfigKey.CustomCloudSaveManifests, result, AppFilePath.UserFile);
+            }
+            catch { }
+            MainWindowViewModel.Instance.AppBarText = "Successfully saved custom Ludusavi Manifests";
+        }
     }
 }